refactor : MVC 패턴 적용 - PostDetailModal 파일 분리
