+++ src/components/modalViews/index.ts
@@ -5,4 +5,4 @@ export { default as TagCreateModal } from "./TagCreateModal";
 export { default as EditNameModal } from "./ProfileModal/EditNameModal";
 export { default as EditPasswordModal } from "./ProfileModal/EditPasswordModal";
 export { default as ChangeImageModal } from "./ProfileModal/ChangeImageModal";
-export { default as PostDetailModal } from "./PostDetailModal/PostDetailModal";
+export { default as PostDetailModal } from "./PostDetailModal/PostDetailModal.view";
